archive/requirements.txt
black==24.4.0
blinker==1.7.0
boto3==1.34.33
botocore==1.34.33
cached-property==1.5.2
click==8.1.7
dataclasses==0.6
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
platformdirs==4.2.0
PyJWT==2.8.0
pymssql==2.2.11
pyodbc==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.25
sqlalchemy-orm==1.2.10
SQLAlchemy-Utils==0.41.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
Werkzeug==3.0.1
